+++ src/backend/oracle-data-api/src/main/java/ca/bc/gov/open/jag/tco/oracledataapi/OracleDataApiApplication.java
@@ -5,12 +5,11 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
-@SpringBootApplication
+@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
 @EnableScheduling
-@EnableJpaAuditing
 public class OracleDataApiApplication {
 
 	private static Logger logger = LoggerFactory.getLogger(OracleDataApiApplication.class);
