+++ src/backend/oracle-data-api/src/main/java/ca/bc/gov/open/jag/tco/oracledataapi/service/EmailHistoryService.java
@@ -2,8 +2,6 @@ package ca.bc.gov.open.jag.tco.oracledataapi.service;
 
 import java.util.List;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
 import javax.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,9 +15,6 @@ public class EmailHistoryService {
 
 	@Autowired
 	EmailHistoryRepository emailHistoryRepository;
-
-	@PersistenceContext
-	private EntityManager entityManager;
 
 	/**
 	 * Retrieves {@link EmailHistory} records by Ticket Number, delegating to CrudRepository
