+++ src/backend/oracle-data-api/src/test/java/ca/bc/gov/open/jag/tco/oracledataapi/repository/impl/h2/JJDisputeRemarkRepositoryImpl.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ca.bc.gov.open.jag.tco.oracledataapi.model.JJDisputeRemark;
 import ca.bc.gov.open.jag.tco.oracledataapi.repository.JJDisputeRemarkRepository;
 
-@ConditionalOnProperty(name = "repository.jjdispute", havingValue = "h2", matchIfMissing = true)
+@ConditionalOnProperty(name = "repository.jjdispute", havingValue = "h2", matchIfMissing = false)
 @Qualifier("jjDisputeRemarkRepository")
 public interface JJDisputeRemarkRepositoryImpl extends JJDisputeRemarkRepository, JpaRepository<JJDisputeRemark, Long> {
